move stuff into flux-alloc/src/vec/mod.rs
