+++ lib/flux-alloc/src/lib.rs
@@ -1,0 +1,5 @@
+// #![no_std]
+#![cfg_attr(flux, feature(allocator_api))]
+
+#[cfg(flux)]
+pub mod vec;
