+++ tests/tests/lib/vec.rs
@@ -6,9 +6,7 @@ use std::{
 
 use flux_rs::extern_spec;
 
-#[extern_spec]
-#[flux::refined_by(len: int)]
-struct Vec<T, A: Allocator = Global>;
+extern crate flux_alloc;
 
 #[extern_spec]
 #[flux::assoc(fn in_bounds(idx: Self, v: T) -> bool)]
@@ -35,24 +33,6 @@ impl<T, I: SliceIndex<[T]>, A: Allocator> IndexMut<I> for Vec<T, A> {
 }
 
 //---------------------------------------------------------------------------------------
-
-#[extern_spec]
-impl<T> Vec<T> {
-    #[flux::sig(fn() -> Vec<T>[0])]
-    fn new() -> Vec<T>;
-}
-
-#[extern_spec]
-impl<T, A: Allocator> Vec<T, A> {
-    #[flux::sig(fn(self: &mut Vec<T, A>[@n], T) ensures self: Vec<T, A>[n+1])]
-    fn push(v: &mut Vec<T, A>, value: T);
-
-    #[flux::sig(fn(&Vec<T, A>[@n]) -> usize[n])]
-    fn len(v: &Vec<T, A>) -> usize;
-
-    #[flux::sig(fn(self: &mut Vec<T, A>[@n]) -> Option<T> ensures self: Vec<T, A>[if n > 0 { n-1 } else { 0 }])]
-    fn pop(&mut self) -> Option<T>;
-}
 
 #[extern_spec]
 impl<'a, T, A: Allocator> IntoIterator for &'a Vec<T, A> {
